Moved the addVolunteerToEvent function to the logic file. Created flash messages in the controllers instead of the logic file. Modified our test suites to match the new changes.
